app/requirements.txt
openstacksdk~=0.61.0
keystoneauth1==4.5.0
matplotlib~=3.5.1
networkx==2.8
pandas==1.4.2
python-neutronclient~=7.8.0
requests~=2.27.1
python-keystoneclient~=4.4.0
python-glanceclient~=3.6.0
argparse~=1.4.0
numpy==1.22.3
httplib2~=0.20.4
gurobipy~=9.5.1
chronicler~=0.3
python-novaclient~=17.7.0
flask==2.1.1
uwsgi~=2.0.20
mysql-connector-python~=8.0.28
ptvsd~=4.3.2
# mininet~=
